MinGW64
  * build MinGW 64bit without requiring -fpermissive.

+++ src/rigcontrol/FreqControl.cxx
@@ -60,7 +60,7 @@ void cbSelectDigit (Fl_Widget *btn, void * nbr)
 {
 
 	Fl_Button *b = (Fl_Button *)btn;
-	int Nbr = (int)(reinterpret_cast<long> (nbr));
+	int Nbr = reinterpret_cast<intptr_t> (nbr);
 
 	cFreqControl *fc = (cFreqControl *)b->parent();
 	if (fc->hrd_buttons) {
